+++ src/rimraf-native.ts
@@ -2,16 +2,23 @@ import { RimrafOptions } from '.'
 import { promises, rmSync } from './fs.js'
 const { rm } = promises
 
-export const rimrafNative = (path: string, opt: RimrafOptions) =>
-  rm(path, {
+export const rimrafNative = async (
+  path: string,
+  opt: RimrafOptions
+): Promise<boolean> => {
+  await rm(path, {
     ...opt,
     force: true,
     recursive: true,
   })
+  return true
+}
 
-export const rimrafNativeSync = (path: string, opt: RimrafOptions) =>
+export const rimrafNativeSync = (path: string, opt: RimrafOptions): boolean => {
   rmSync(path, {
     ...opt,
     force: true,
     recursive: true,
   })
+  return true
+}