+++ tap-snapshots/test/rimraf-posix.js.test.cjs
@@ -1,0 +1,76 @@
+/* IMPORTANT
+ * This snapshot file is auto-generated, but designed for humans.
+ * It should be checked into source control and tracked carefully.
+ * Re-generate by setting TAP_SNAPSHOT=1 and running tests.
+ * Make sure to inspect the output below.  Do not ignore changes!
+ */
+'use strict'
+exports[`test/rimraf-posix.js TAP filter function filter=i async > paths seen 1`] = `
+Array [
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/a",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/b",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/d",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/e",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/f/g",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/f/h",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/f/i",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/f/i/j",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/f/i/k",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/f/i/l",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/f/i/m",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/f/i/m/n",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-async/c/f/i/m/o",
+]
+`
+
+exports[`test/rimraf-posix.js TAP filter function filter=i sync > paths seen 1`] = `
+Array [
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/a",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/b",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/d",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/e",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/f/g",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/f/h",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/f/i",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/f/i/j",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/f/i/k",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/f/i/l",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/f/i/m",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/f/i/m/n",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-i-sync/c/f/i/m/o",
+]
+`
+
+exports[`test/rimraf-posix.js TAP filter function filter=j async > paths seen 1`] = `
+Array [
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/a",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/b",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/c/d",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/c/e",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/c/f/g",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/c/f/h",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/c/f/i/j",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/c/f/i/k",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/c/f/i/l",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/c/f/i/m",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/c/f/i/m/n",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-async/c/f/i/m/o",
+]
+`
+
+exports[`test/rimraf-posix.js TAP filter function filter=j sync > paths seen 1`] = `
+Array [
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/a",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/b",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/c/d",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/c/e",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/c/f/g",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/c/f/h",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/c/f/i/j",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/c/f/i/k",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/c/f/i/l",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/c/f/i/m",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/c/f/i/m/n",
+  "test/tap-testdir-rimraf-posix-filter-function-filter-j-sync/c/f/i/m/o",
+]
+`